requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.2.0=pypi_0
blas=1.0=openblas
bottleneck=1.3.5=py311ha0d4635_0
brotlipy=0.7.0=py311h80987f9_1002
bzip2=1.0.8=h620ffc9_4
ca-certificates=2022.12.7=h4653dfc_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py311h80987f9_3
charset-normalizer=3.1.0=pypi_0
click=8.0.3=pyhd3eb1b0_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.0.7=pypi_0
cryptography=39.0.1=py311h834c97f_0
cycler=0.11.0=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
dnspython=2.3.0=pypi_0
et_xmlfile=1.1.0=pyhd8ed1ab_0
flask=2.0.3=pyhd3eb1b0_0
flask-cors=3.0.10=pyhd3eb1b0_0
fonttools=4.39.2=pypi_0
frozenlist=1.3.3=pypi_0
idna=3.4=pyhd8ed1ab_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.2.0=pypi_0
kiwisolver=1.4.4=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.2=hca03da5_6
libgfortran=5.0.0=11_2_0_he6877d6_26
libgfortran5=11.2.0=he6877d6_26
libopenblas=0.3.20=hea475bc_0
llvm-openmp=12.0.0=haf9daa7_1
loguru=0.6.0=pypi_0
markupsafe=2.1.1=py311h80987f9_0
matplotlib=3.7.1=pypi_0
more-itertools=9.1.0=pypi_0
multidict=6.0.4=pypi_0
ncurses=6.4=h313beb8_0
numexpr=2.8.4=py311hdeb1df9_0
numpy=1.24.2=pypi_0
numpy-base=1.23.5=py311h9eb1c70_0
openai=0.27.2=pypi_0
openpyxl=3.0.9=pyhd8ed1ab_0
openssl=1.1.1t=h03a7124_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.3=py311h6956b77_0
pandas-stubs=1.5.3.230321=pyhd8ed1ab_0
pillow=9.4.0=pypi_0
pinecone-client=2.2.1=pypi_0
pip=23.0.1=py311hca03da5_0
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=23.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=pyha2e5f31_6
python=3.11.0=hc0d8a6c_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=1.0.0=pypi_0
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
readline=8.2=h1a28f6b_0
requests=2.28.2=pyhd8ed1ab_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=65.6.3=py311hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.1=h80987f9_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=hb8d0fd4_0
tqdm=4.65.0=pyhd8ed1ab_1
types-pytz=2022.7.1.2=pyhd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.15=pyhd8ed1ab_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.38.4=py311hca03da5_0
xz=5.2.10=h80987f9_1
yarl=1.8.2=pypi_0
zlib=1.2.13=h5a0b063_0
